Extracting common data generators
Leveraging a base class to generate most of the behavior.
