+++ spec/data_generators/sipity/data_generators/base/submission_window_processing_generator_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
-require 'sipity/data_generators/etd/submission_window_processing_generator'
+require 'sipity/data_generators/base/submission_window_processing_generator'
 
 module Sipity
   module DataGenerators
-    module Etd
+    module Base
       # Responsible for generating the submission window for the ETD work area.
       RSpec.describe SubmissionWindowProcessingGenerator do
         let(:work_area) { Models::WorkArea.new(id: 1, slug: 'etd') }
